[google_maps_flutter] Fix async handling in examples (#4729)
Applies the fixes from https://github.com/flutter/packages/pull/4171 to the minimal ExampleGoogleMaps class used in the Android and iOS implementation package examples, as they had the same issue.

Also sets up unit testing of the examples in order to add testing of this fix. The fake is copied directly from the app-facing package, and the unit tests are slightly modified versions of the tests in the app-facing package as well.