Fix Google Maps rendering issues in TLHC mode when using LATEST renderer (#5408)
The Google Maps LATEST renderer always uses a TextureView. We can use a
signal from this TextureView to perform view invalidation that fixes the
rendering glitches (missing updates) in TLHC mode.

NOTE: We have an internal bug 311013682 requesting an official way of
achieving this functionality but if the bug is ever acted on it will
take many months/years before we can rely on this functionality.

In the meantime, chain the internal SurfaceTextureListener with our own
and piggyback on the OnSurfaceTextureUpdated callback to invalidate the
MapView.

Fixes https://github.com/flutter/flutter/issues/103686

Tested on an emulator and a physical device (Pixel 6 Pro).