[various] Bump androidx.test:core to 1.4.0 (#8710)
This is pulled out of https://github.com/flutter/packages/pull/8693 to land on its own while the other issues with that PR are investigated, since it's a useful change regardless.

This updates all packages in the repo to have `androidx.test:core` 1.4.0 if it's not already higher. This fixes a latent problem where older versions of `androidx.test.core` aren't compatible with targeting the current SDK, because it has activities that are not annotated as exported. This issue is currently being masked by the dependency on `integration_test`, which forced a higher version of that package, but `integration_test` will stop exporting that dependency onto the app as of recent master, dropping all the packages here back to a too-old `androidx.test:core`.